@@ -3,39 +3,36 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^5.0.3",
-    "eslint-plugin-react": "^7.0.1",
-    "react-scripts": "1.0.7",
-    "redux-devtools": "^3.4.0"
+    "eslint": "^4.17.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.9.1",
+    "react-scripts": "1.1.4",
+    "redux-devtools": "^3.4.1"
   },
   "dependencies": {
-    "google-maps-react": "^1.0.19",
-    "history": "^4.6.1",
-    "invariant": "^2.2.2",
-    "lodash": "^4.17.4",
-    "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-ga": "^2.2.0",
-    "react-helmet": "^5.1.3",
-    "react-image-gallery": "^0.8.3",
-    "react-redux": "^5.0.4",
-    "react-redux-form": "^1.12.1",
-    "react-redux-toastr": "^7.1.5",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
+    "history": "^4.7.2",
+    "lodash": "^4.17.10",
+    "prop-types": "^15.6.1",
+    "react": "^16.4.1",
+    "react-detect-offline": "^2.1.1",
+    "react-dom": "^16.4.1",
+    "react-ga": "^2.5.3",
+    "react-helmet": "^5.2.0",
+    "react-image-gallery": "^0.8.9",
+    "react-redux": "^5.0.7",
+    "react-redux-toastr": "^7.3.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "react-router-redux": "^4.0.8",
-    "react-share": "^1.16.0",
-    "redux": "^3.6.0",
+    "react-share": "^2.2.0",
+    "redux": "^4.0.0",
     "redux-logger": "^3.0.6",
-    "redux-persist": "^4.10.1",
-    "redux-polyglot": "^0.5.3",
-    "redux-thunk": "^2.2.0",
-    "semantic-ui-react": "^0.72.0",
-    "whatwg-fetch": "^2.0.3"
+    "redux-persist": "^5.10.0",
+    "redux-thunk": "^2.3.0",
+    "semantic-ui-react": "^0.81.1",
+    "whatwg-fetch": "^2.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
