@@ -9,7 +9,17 @@ import { getReviewsFetching, getReviews, reviewPropType } from './reducer';
 import Review from './Review';
 
 class Reviews extends Component {
-  componentWillMount() {
+  componentDidMount() {
+    this.readReviews();
+  }
+
+  componentDidUpdate(prevProps) {
+    if (this.props.productId !== prevProps.productId) {
+      this.readReviews();
+    }
+  }
+
+  readReviews() {
     const { dispatch } = this.props;
     dispatch(fetchReviews(this.props.productId));
   }
@@ -35,9 +45,9 @@ Reviews.propTypes = {
   reviews: PropTypes.arrayOf(reviewPropType).isRequired,
 };
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state, props) => ({
   loading: getReviewsFetching(state.reviews),
-  reviews: getReviews(state.reviews),
+  reviews: getReviews(state.reviews, props.productId),
 });
 
 function mapDispatchToProps(dispatch) {
