@@ -1,71 +1,85 @@
-import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
-import { createTransform, persistStore, autoRehydrate } from 'redux-persist';
+import { createStore, applyMiddleware } from 'redux';
+import { persistCombineReducers, persistReducer, persistStore } from 'redux-persist';
+import storage from 'redux-persist/lib/storage'; // defaults to localStorage for web
 import { reducer as toastrReducer } from 'react-redux-toastr';
 import thunk from 'redux-thunk';
 // import logger from 'redux-logger';
 import createHistory from 'history/createBrowserHistory';
 import { routerMiddleware } from 'react-router-redux';
-import sideMenuVisible from './components/NavBar/reducer';
 
+import config from './config/config';
 import categories from './views/Categories/reducer';
 import products from './views/Products/reducer';
 import reviews from './components/Reviews/reducer';
 import cart from './views/Cart/reducer';
 import variations from './components/Variations/reducer';
+import sideMenuVisible from './components/NavBar/reducer';
+
+const rootPersistConfig = {
+  key: 'root',
+  storage,
+  blacklist: [
+    'sideMenuVisible',
+    'toastr',
+  ],
+  // debug: true,
+};
+
+const rootReducer = persistCombineReducers(rootPersistConfig, {
+  categories: persistReducer(
+    {
+      key: 'categories',
+      storage,
+      blacklist: ['isFetching'],
+    },
+    categories,
+  ),
+  products: persistReducer(
+    {
+      key: 'products',
+      storage,
+      blacklist: ['isFetching'],
+    },
+    products,
+  ),
+  reviews: persistReducer(
+    {
+      key: 'reviews',
+      storage,
+      blacklist: ['isFetching'],
+    },
+    reviews,
+  ),
+  variations: persistReducer(
+    {
+      key: 'variations',
+      storage,
+      blacklist: ['isFetching'],
+    },
+    variations,
+  ),
+  cart,
+  sideMenuVisible,
+  toastr: persistReducer(
+    {
+      key: 'toastr',
+      storage,
+    },
+    toastrReducer,
+  ),
+});
 
 const history = createHistory();
 
-const defaultState = {
-  sideMenuVisible: false,
-  categories: {
-    items: [],
-    isFetching: 0,
-  },
-  products: {
-    items: [],
-    isFetching: 0,
-  },
-  reviews: {
-    items: [],
-    isFetching: 0,
-  },
-  variations: {
-    items: [],
-    isFetching: 0,
-  },
-  cart: {
-    items: [],
-  },
-};
-
-const rootReducer = combineReducers({
-  sideMenuVisible,
-  categories,
-  products,
-  reviews,
-  cart,
-  variations,
-  toastr: toastrReducer,
-});
-
-const skipIsFetchingTransform = createTransform((inboundState, key) => {
-  if (key !== 'products' && key !== 'categories' && key !== 'reviews' && key !== 'variations') { return inboundState; }
-  return {
-    ...inboundState,
-    isFetching: undefined,
-  };
-});
-
 const store = createStore(
   rootReducer,
-  defaultState,
-  compose(applyMiddleware(thunk, routerMiddleware(history)), autoRehydrate()),
+  undefined,
+  applyMiddleware(thunk, routerMiddleware(history)),
 );
 
-persistStore(store, {
-  blacklist: ['sideMenuVisible', 'toastr', 'reviews'],
-  transforms: [skipIsFetchingTransform],
-});
+if (config.OFFLINE) {
+  persistStore(store);
+}
 
 export { history };
 export default store;
