@@ -11,7 +11,9 @@ class CategoryCard extends Component {
     return (
       <Card centered>
         <Card.Content>
-          {this.props.src !== '' ? <CircularImage src={this.props.src} /> : null}
+          {this.props.src !== '' ? (
+            <CircularImage src={this.props.src} />
+          ) : null}
           <Segment basic className="category-meta-container">
             <Card.Header as={Header} className="break-words">
               {this.props.name}
