@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import { ShareButtons } from 'react-share';
+import { FacebookShareButton, GooglePlusShareButton, TwitterShareButton } from 'react-share';
 import { Header, Card, Icon, Grid } from 'semantic-ui-react';
 
 class SocialBox extends Component {
@@ -14,22 +14,22 @@ class SocialBox extends Component {
           <Grid doubling centered>
             <Grid.Row>
               <Grid.Column width={4}>
-                <ShareButtons.FacebookShareButton className="no-outline" url={this.props.permalink}>
+                <FacebookShareButton className="no-outline" url={this.props.permalink}>
                   <Icon name="facebook f" color="purple" size="big" />
-                </ShareButtons.FacebookShareButton>
+                </FacebookShareButton>
               </Grid.Column>
               <Grid.Column width={4}>
-                <ShareButtons.GooglePlusShareButton
+                <GooglePlusShareButton
                   className="no-outline"
                   url={this.props.permalink}
                 >
                   <Icon name="google plus" color="purple" size="big" />
-                </ShareButtons.GooglePlusShareButton>
+                </GooglePlusShareButton>
               </Grid.Column>
               <Grid.Column width={4}>
-                <ShareButtons.TwitterShareButton className="no-outline" url={this.props.permalink}>
+                <TwitterShareButton className="no-outline" url={this.props.permalink}>
                   <Icon name="twitter" color="purple" size="big" />
-                </ShareButtons.TwitterShareButton>
+                </TwitterShareButton>
               </Grid.Column>
             </Grid.Row>
           </Grid>
